Added initial session management
A unique username is now required for each client that try to connect to the server. The server handles the requests and registers the new user if his username is not already in his list of connected users. Test section of both client and server has been modified using a "chat" message type. So actually, a simple basic chat system has also been implemented with this commit.