Timing system working
Many changes to address many bugs, but now the system will correctly start the internal timer when the first client will connect to the server. When new users will connect, the server will correctly send the current state of the timer. Clients' windows are updated accordingly to show the remaining time.