Configured handlers for players movement
Maps classes are used only for the graphical aspect of the scenery (event listeners and so on). SceneryManager has been removed since it was redundant; it has been substituted by Scenery class and its subclasses. Now Scenery class has not only a graph but also a hashmap. The hashmap contains strings as key and places as value. This is used in order to catch user input and convert them to the appropriate corresponding Place object for players movements routines.