Fixed ugly player stuck in place bug
The bug was caused by an improper use of a JGraphT library method. To avoid future hassles, a new method has been added to quickly and reliably find the destination of a path, and old methods have been marked as deprecated. An overridden equals has been added to Place, since it is used by the new getPathTarget() method.